# src/main/resources/swagger/api.yml
openapi: 3.0.1
info:
  title: Studentpoint API
  description: Studentpoint API documentation
  version: 0.0.1
servers:
  - url: http://127.0.0.1:8080/api
...
paths:
...
  /conduct/get:
    get:
      tags:
        - classes-resource
      operationId: conductGet
      parameters:
        - name: classIName
          in: query
          required: true
          schema:
            type: string
        - name: academicYear
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConductScoresCustomDTO'
  /conduct/create:
    post:
      tags:
        - classes-resource
      operationId: conductCreate
      parameters:
        - name: classIName
          in: query
          required: true
          schema:
            type: string
        - name: academicYear
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseObject'
...
components:
  schemas:
...
    ConductScoresCustomDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        conductScoresCode:
          type: string
        academicYear:
          type: string
        score:
          type: integer
          format: int32
        classification:
          type: integer
          format: int32
        evaluation:
          type: string
        createdBy:
          type: string
        createdDate:
          type: string
        lastModifiedBy:
          type: string
        lastModifiedDate:
          type: string
        student:
          $ref: '#/components/schemas/StudentCustomDTO'
    StudentCustomDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        studentCode:
          type: string
        fullName:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        address:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        notes:
          type: string
        status:
          type: string
        dateEnrollment:
          type: string
        createdBy:
          type: string
        createdDate:
          type: string
        lastModifiedBy:
          type: string
        lastModifiedDate:
          type: string
        className:
          type: string
        courseYear:
          type: string
        faculties:
          $ref: '#/components/schemas/FacultiesCustomDTO'
...
    FacultiesCustomDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        facultyCode:
          type: string
        facultyName:
          type: string
        description:
          type: string
        establishedDate:
          type: string
        phoneNumber:
          type: string
        location:
          type: string
        notes:
          type: string
        parentId:
          type: string
        createdBy:
          type: string
        createdDate:
          type: string
        lastModifiedBy:
          type: string
        lastModifiedDate:
          type: string
...
    ResponseObject:
      type: object
      additionalProperties: true
      example:
        someKey: 'value'
        anotherKey: 123
...